git-svn-id: svn+ssh://svn.code.sf.net/p/dynamicdocs/code@129 05742d63-fc59-4ccd-b3f0-d7e2a2a8a88d

+++ ddocclient/kksfactory/KKSViewFactory.cpp
@@ -701,7 +701,7 @@ void KKSViewFactory :: loadEIOEx (QWidget *editor,
  */
 void KKSViewFactory :: updateEIOEx (KKSLoader *l,
                                     KKSObject *pObj,
-                                    int idObjEx,
+                                    qint64 idObjEx,
                                     const KKSTemplate *t,
                                     QAbstractItemModel *sourceModel,
                                     int row,
